requirements.txt
asgiref==3.7.2
axios==0.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
idna==3.6
lxml==5.1.0
markdown-it-py==3.0.0
mdurl==0.1.2
openpyxl==3.1.2
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pytz==2024.1
requests==2.31.0
rich==13.7.1
sqlparse==0.4.4
urllib3==2.2.1
